Update Screen.java
Added the space KeyEvent that produces the shot, although it doesn't work. I tested the code that would produce the shot using the addMovingObjects method, and it works, so it's something about the KeyListener that I'm not understand, so I'm going to leave it up to you, Steven, to fix this bug.